Opinionated changes: namespace and remov CDATA
Namespace: when exporting a lot of attributes prefixing everything with
`assvg` makes for a lot work later when trying to use them. It's all
data so just call them "d". (As a bonus we sidestep associating all as
"ass"; probably only relevant for english-as-1st-language folk.)

CDATA: just makes things messy. Besides "You should almost never need to
use CDATA Sections. The CDATA mechanism was designed to let an author
quote fragments of text containing markup characters (the
open-angle-bracket and the ampersand)"
  -- http://xml.silmaril.ie/cdata.html